+++ CHANGELOG.md
@@ -939,5 +939,6 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Allow CustomCard to have different card art in SCV on upgrade (kiooeht/Vex)
 
 #### dev ####
+* Dear ImGui (kiooeht)
 * Fix black screen when using post-processing effects and cursor is hidden (kiooeht)
 * Don't use energy when infinite energy is enabled (kiooeht)